Looks like the task quantity computations need more work, but at least
there's now a test of nullspace projections that does not
crash. There's one task for all odd-numbered joints, another one for
all even-numbered joints, and when they are placed in a hierarchy they
should accumulate to the same torques as a task that does all the
joints (and compensates for mass-inertia effects). This last check is
not yet done though.